requirements.txt
flask
selenium
webdriver-manager
python-dotenv
schedule
Flask-Mail
